LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-remove-head-expected.txt


PASS Meta color-scheme applies.
FAIL Initial value after removing head including meta color-scheme. assert_equals: expected "light" but got "dark"
